refactor class filenames to CamelCase
